+++ src/Repository/StaffRepository.php
@@ -1,0 +1,25 @@
+<?php
+
+/*
+ * This script is part of incwadi/core
+ */
+
+namespace Incwadi\Core\Repository;
+
+use Doctrine\Bundle\DoctrineBundle\Repository\ServiceEntityRepository;
+use Doctrine\Common\Persistence\ManagerRegistry;
+use Incwadi\Core\Entity\Staff;
+
+/**
+ * @method Staff|null find($id, $lockMode = null, $lockVersion = null)
+ * @method Staff|null findOneBy(array $criteria, array $orderBy = null)
+ * @method Staff[]    findAll()
+ * @method Staff[]    findBy(array $criteria, array $orderBy = null, $limit = null, $offset = null)
+ */
+class StaffRepository extends ServiceEntityRepository
+{
+    public function __construct(ManagerRegistry $registry)
+    {
+        parent::__construct($registry, Staff::class);
+    }
+}