+++ CHANGELOG.md
@@ -3,7 +3,8 @@
 ## 1.1.0
 
 - feature: add info whether the user has role user and admin
-
+- feature: rename customer to staff
+-
 ## 1.0.0
 
 - First Release